change file names to make it compatible with latest merge
