+++ packages/react-app/pages/admin/batches.jsx
@@ -1,11 +1,14 @@
-import React, { useEffect, useState, useMemo, useCallback } from "react";
+import React, { useEffect, useState, useMemo } from "react";
 import axios from "axios";
+import moment from "moment";
+import NextLink from "next/link";
 import {
   Box,
   Button,
   ButtonGroup,
   Center,
   Container,
+  Link,
   Text,
   Table,
   Thead,
@@ -17,28 +20,78 @@ import {
   Flex,
   Select,
   Tooltip,
+  HStack,
   Input,
   InputRightElement,
   InputGroup,
-  IconButton,
+  useClipboard,
 } from "@chakra-ui/react";
-
-import BatchNumberCell from "../../components/batches/BatchNumberCell";
-import { SearchIcon, TriangleDownIcon, TriangleUpIcon, EditIcon, AddIcon } from "@chakra-ui/icons";
 import { useTable, usePagination, useSortBy, useFilters } from "react-table";
-import useCustomColorModes from "../../hooks/useCustomColorModes";
-import BatchLinksCell from "../../components/batches/BatchLinksCell";
-import BatchStatusCell from "../../components/batches/BatchStatusCell";
-import ExactDateWithTooltip from "../../components/batches/ExactDateWithTooltip";
+import { CopyIcon, SearchIcon, TriangleDownIcon, TriangleUpIcon } from "@chakra-ui/icons";
+import BuilderListSkeleton from "../../components/skeletons/BuilderListSkeleton";
+import DateWithTooltip from "../../components/DateWithTooltip";
+import SocialLink from "../../components/SocialLink";
+import Address from "../../components/Address";
+import { bySocialWeight } from "../../data/socials";
 import { USER_ROLES } from "../../helpers/constants";
-import { BatchCrudFormModal } from "../../components/batches/BatchCrudForm";
-import BatchesListSkeleton from "../../components/skeletons/BatchesListSkeleton";
+import StreamTableCell from "../../components/StreamTableCell";
+import BuilderFlags from "../../components/builder/BuilderFlags";
+import useCustomColorModes from "../../hooks/useCustomColorModes";
+import BatchColumnFilter from "../../components/BatchColumnFilter";
 
-const serverPathBatches = "/builders/batches";
-const serverPathBatchGraduateBuilders = "/builders/batchGraduateBuilders";
+const serverPath = "/builders/batchBuilders";
 
-// TODO: double check this, adapt to batch number
-const BatchColumnFilter = ({ column: { filterValue, setFilter } }) => {
+const builderCreated = builder => {
+  return builder?.creationTimestamp || 0;
+};
+
+const BuilderSocialLinksCell = ({ builder, isAdmin }) => {
+  const socials = Object.entries(builder.socialLinks ?? {}).sort(bySocialWeight);
+
+  return (
+    <Flex direction="column">
+      <HStack spacing={3} alignItems="center" justifyContent="center">
+        {socials.length ? (
+          socials.map(([socialId, socialValue]) => <SocialLink id={socialId} key={socialId} value={socialValue} />)
+        ) : (
+          <Box>-</Box>
+        )}
+      </HStack>
+      {isAdmin && <BuilderFlags builder={builder} />}
+    </Flex>
+  );
+};
+
+const BuilderAddressCell = ({ builder, mainnetProvider }) => {
+  const { hasCopied, onCopy } = useClipboard(builder?.id);
+
+  return (
+    <Flex alignItems="center" whiteSpace="nowrap">
+      <NextLink href={`/builders/${builder.id}`} passHref>
+        <Link pos="relative" d="inline-block">
+          <Address address={builder.id} ensProvider={mainnetProvider} w="12.5" fontSize="16" cachedEns={builder.ens} />
+        </Link>
+      </NextLink>
+      <Tooltip label={hasCopied ? "Copied!" : "Copy address"} closeOnClick={false}>
+        <CopyIcon cursor="pointer" onClick={onCopy} ml={"7px"} />
+      </Tooltip>
+    </Flex>
+  );
+};
+
+const BuilderStatusCell = ({ status }) => {
+  return (
+    <Tooltip label={moment(status?.timestamp).fromNow()}>
+      <Text maxW="350">{status?.text}</Text>
+    </Tooltip>
+  );
+};
+
+const BuilderBuildsCell = ({ buildCount }) => {
+  return <Text>{buildCount}</Text>;
+};
+
+const EnsColumnFilter = ({ column: { filterValue, setFilter } }) => {
   const { baseColor } = useCustomColorModes();
 
   return (
@@ -48,169 +101,165 @@ const BatchColumnFilter = ({ column: { filterValue, setFilter } }) => {
       onChange={e => {
         setFilter(e.target.value || undefined);
       }}
-      placeholder="Search for batch"
+      placeholder="Search builder"
       bgColor={baseColor}
       mb={8}
     />
   );
 };
 
-export default function Batches({ serverUrl, userRole, mainnetProvider }) {
-  const [batches, setBatches] = useState([]);
-  const [isLoadingBatches, setIsLoadingBatches] = useState(false);
-  const { baseColor } = useCustomColorModes();
+const isValueOnEnsOrSocials = (builder, filterValue) => {
+  const isOnEns = String(builder.ens).toLowerCase().includes(String(filterValue).toLowerCase());
+  const isOnSocials =
+    builder.socialLinks &&
+    Object.entries(builder.socialLinks ?? {}).some(([_, social]) => {
+      return String(social).toLowerCase().includes(String(filterValue).toLowerCase());
+    });
+
+  return isOnEns || isOnSocials;
+};
+
+const isInBatch = (builder, filterValue) => {
+  if (!builder.batch?.number) return false;
+
+  return builder.batch?.number === filterValue;
+};
+
+export default function BuilderListView({ serverUrl, mainnetProvider, userRole }) {
+  const [builders, setBuilders] = useState([]);
+  const [amountBuilders, setAmountBuilders] = useState(0);
+  const [isLoadingBuilders, setIsLoadingBuilders] = useState(false);
   const isAdmin = userRole === USER_ROLES.admin;
-  const [isEditModalOpen, setIsEditModalOpen] = useState(false);
-  const [selectedBatch, setSelectedBatch] = useState(null);
-  const [amountBatches, setAmountBatches] = useState();
-  const [graduatesCount, setGraduatesCount] = useState({});
+  const isLoggedIn = userRole !== null && userRole !== USER_ROLES.anonymous;
+  const { baseColor } = useCustomColorModes();
 
-  const [isAddModalOpen, setIsAddModalOpen] = useState(false);
+  const ensFiltering = (rows, id, filterValue) => {
+    if (filterValue.length < 3) {
+      return rows;
+    }
 
-  const batchFiltering = (rows, id, filterValue) => {
     return rows.filter(row => {
       const rowValue = row.values[id];
-      const rowValueString = String(rowValue).toLowerCase();
-      const filterValueString = String(filterValue).toLowerCase();
-      return rowValueString.includes(filterValueString);
+      return rowValue !== undefined ? isValueOnEnsOrSocials(rowValue, filterValue) : true;
     });
   };
 
-  const fetchBatches = useCallback(async () => {
-    setIsLoadingBatches(true);
-    try {
-      const fetchedBatches = await axios.get(serverUrl + serverPathBatches);
-      const processedBatches = fetchedBatches.data.map(batch => ({
-        batch: batch,
-        batchNumber: batch.number,
-        status: batch.status,
-        startDate: batch.startDate,
-      }));
-      setBatches(processedBatches);
-      setAmountBatches(processedBatches.length);
-    } catch (error) {
-      console.error("Error fetching batches:", error);
-    } finally {
-      setIsLoadingBatches(false);
+  const batchFiltering = (rows, id, filterValue) => {
+    if (filterValue === "allBatches") {
+      return rows.filter(row => row.values[id]);
     }
+
+    return rows.filter(row => {
+      const rowValue = row.values[id];
+      return rowValue !== undefined ? isInBatch(rowValue, filterValue) : true;
+    });
+  };
+
+  useEffect(() => {
+    async function fetchBuilders() {
+      setIsLoadingBuilders(true);
+      const fetchedBuilders = await axios.get(serverUrl + serverPath);
+
+      const processedBuilders = fetchedBuilders.data
+        .filter(builder => builder.ens !== "austingriffith.eth")
+        .map(builder => ({
+          builder,
+          status: builder.status,
+          stream: builder,
+          builds: builder.builds?.length || 0,
+          socials: builder,
+          userCreated: builderCreated(builder),
+        }));
+
+      setBuilders(processedBuilders);
+      setIsLoadingBuilders(false);
+    }
+
+    fetchBuilders();
   }, [serverUrl]);
 
-  const fetchGraduatesCount = useCallback(async () => {
-    const fetchedBatchGraduateBuilders = await axios.get(serverUrl + serverPathBatchGraduateBuilders);
-    const graduatesCounts = {};
-
-    batches.forEach(
-      batch => {
-        const batchNumber = batch.batchNumber;
-        graduatesCounts[batchNumber] = 0;
-
-        fetchedBatchGraduateBuilders.data.forEach(builder => {
-          if (builder.batch && builder.batch.number === String(batchNumber)) {
-            graduatesCounts[batchNumber]++;
-          }
-        });
-      },
-      [batches],
+  const BatchFilterComponent = useMemo(() => {
+    const Component = ({ column }) => (
+      <BatchColumnFilter filterValue={column.filterValue} setFilter={column.setFilter} builders={builders} />
     );
+    Component.displayName = "BatchFilterComponent";
+    return Component;
+  }, [builders]);
 
-    setGraduatesCount(graduatesCounts);
-  }, [batches, serverUrl]);
-
-  useEffect(() => {
-    fetchBatches();
-  }, [fetchBatches]);
-
-  useEffect(() => {
-    if (batches.length > 0) {
-      fetchGraduatesCount();
-    }
-  }, [batches, fetchGraduatesCount]);
-
-  const BatchNumberCellComponent = ({ row }) => (
-    <BatchNumberCell batch={row.original.batchNumber} status={row.original.status} />
+  const BuilderAddressCellComponent = ({ value }) => (
+    <BuilderAddressCell builder={value} mainnetProvider={mainnetProvider} />
   );
-  const BatchCreatedCellComponent = ({ value }) => {
-    return <ExactDateWithTooltip timestamp={value} />;
+  const BuilderStatusCellComponent = ({ value }) => <BuilderStatusCell status={value} />;
+  const BuilderBuildsCellComponent = ({ value }) => <BuilderBuildsCell buildCount={value} />;
+  const StreamTableCellComponent = ({ value }) => {
+    return value?.graduated?.status ? "" : <StreamTableCell builder={value} />;
   };
-  const BatchLinksCellComponent = ({ value }) => <BatchLinksCell batch={value} />;
-  const BatchStatusCellComponent = ({ value }) => <BatchStatusCell status={value} />;
-  const BatchGraduatesCellComponent = ({ row }) => {
-    const count = graduatesCount[row.original.batchNumber] || 0;
-    return <Text>{count}</Text>;
+  const BuilderSocialLinksCellComponent = ({ value }) => <BuilderSocialLinksCell builder={value} isAdmin={isAdmin} />;
+  const UserCreatedCellComponent = ({ value }) => {
+    return <DateWithTooltip timestamp={value} />;
   };
-
-  const BatchEditComponent = ({ row }) => (
-    <Tooltip label="Edit Batch">
-      <IconButton
-        aria-label="Edit batch"
-        icon={<EditIcon />}
-        size="sm"
-        variant="ghost"
-        isDisabled={!isAdmin}
-        onClick={() => {
-          setSelectedBatch({
-            ...row.original.batch,
-            id: row.original.batch.id,
-          });
-          setIsEditModalOpen(true);
-        }}
-      />
-    </Tooltip>
-  );
 
   const columns = useMemo(
     () => {
       const allColumns = [
         {
-          Header: "Batch",
-          accessor: "batchNumber",
+          Header: "Builder",
+          accessor: "builder",
+          disableSortBy: true,
           canFilter: true,
-          Filter: BatchColumnFilter,
-          filter: batchFiltering,
-          Cell: BatchNumberCellComponent,
+          Filter: EnsColumnFilter,
+          filter: ensFiltering,
+          Cell: BuilderAddressCellComponent,
         },
         {
           Header: "Status",
           accessor: "status",
           disableSortBy: true,
           disableFilters: true,
-          Cell: BatchStatusCellComponent,
+          Cell: BuilderStatusCellComponent,
         },
         {
-          Header: "Start Date",
-          accessor: "startDate",
+          Header: "Builds",
+          accessor: "builds",
+          sortDescFirst: true,
           disableFilters: true,
-          Cell: BatchCreatedCellComponent,
+          Cell: BuilderBuildsCellComponent,
         },
         {
-          Header: "Graduates",
-          accessor: row => graduatesCount[row.batchNumber] || 0,
+          Header: "Batch",
+          accessor: "stream",
           disableFilters: true,
-          Cell: BatchGraduatesCellComponent,
+          Filter: BatchFilterComponent,
+          filter: batchFiltering,
+          // // Sorting by stream cap for now.
+          // sortType: (rowA, rowB) =>
+          //   Number(rowA.values?.stream?.cap || 0) < Number(rowB.values?.stream?.cap || 0) ? 1 : -1,
+          Cell: StreamTableCellComponent,
         },
         {
-          Header: "Links",
-          accessor: "batch",
+          Header: "Socials",
+          accessor: "socials",
           disableSortBy: true,
           disableFilters: true,
-          Cell: BatchLinksCellComponent,
+          Cell: BuilderSocialLinksCellComponent,
         },
         {
-          Header: "Edit",
-          disableSortBy: true,
+          Header: "User Created",
+          accessor: "userCreated",
+          sortAscFirst: true,
           disableFilters: true,
-          Cell: BatchEditComponent,
+          Cell: UserCreatedCellComponent,
         },
       ];
 
-      if (!isAdmin) {
-        allColumns.splice(5, 1);
+      if (!isLoggedIn) {
+        allColumns.splice(4, 1);
       }
 
       return allColumns;
     },
     // eslint-disable-next-line
-    [userRole, graduatesCount],
+    [userRole, BatchFilterComponent, builders],
   );
 
   const {
@@ -231,151 +280,127 @@ export default function Batches({ serverUrl, userRole, mainnetProvider }) {
   } = useTable(
     {
       columns,
-      data: batches,
-      initialState: { pageIndex: 0, pageSize: 25, sortBy: useMemo(() => [{ id: "batchNumber", desc: true }], []) },
+      data: builders,
+      initialState: { pageIndex: 0, pageSize: 25, sortBy: useMemo(() => [{ id: "stream", desc: true }], []) },
     },
     useFilters,
     useSortBy,
     usePagination,
   );
 
-  const batchFilter = headerGroups[0].headers[0];
+  useEffect(() => {
+    setAmountBuilders(page.length);
+  }, [page]);
+
+  const ensFilter = headerGroups[0].headers[0];
+  const batchFilter = headerGroups[0].headers[3];
 
   return (
-    <>
-      <Container maxW="container.xl">
-        {isLoadingBatches ? (
-          <BatchesListSkeleton />
-        ) : (
-          <Box overflowX={{ base: "auto", lg: "visible" }} mb={8}>
-            <Center mb={5} flexDir="column">
-              <Box mb={2}>
-                <chakra.strong mr={2}>Total batches:</chakra.strong>
-                {amountBatches}
+    <Container maxW="container.xl">
+      {isLoadingBuilders ? (
+        <BuilderListSkeleton />
+      ) : (
+        <Box overflowX={{ base: "auto", lg: "visible" }} mb={8}>
+          <Center mb={5} flexDir="column">
+            <Box mb={2}>
+              <chakra.strong mr={2}>Total builders:</chakra.strong>
+              {amountBuilders}
+            </Box>
+            <Flex direction={{ base: "column", md: "row" }} alignItems="center" mb={4}>
+              <InputGroup mr={{ md: 4 }} mb={{ base: 4, md: 0 }} width={{ base: "100%", md: "auto" }} height="40px">
+                {ensFilter.render("Filter")}
+                <InputRightElement pointerEvents="none" color="gray.300" fontSize="1.2em" children={<SearchIcon />} />
+              </InputGroup>
+
+              <Box width={{ base: "100%", md: "auto" }} height="40px">
+                {batchFilter.render("Filter")}
               </Box>
-              <Flex direction={{ base: "column", md: "row" }} alignItems="center" mb={4}>
-                <InputGroup mr={{ md: 4 }} mb={{ base: 4, md: 0 }} width={{ base: "100%", md: "auto" }} height="40px">
-                  {batchFilter.render("Filter")}
-                  <InputRightElement pointerEvents="none" color="gray.300" fontSize="1.2em" children={<SearchIcon />} />
-                </InputGroup>
-                <Button
-                  leftIcon={<AddIcon />}
-                  isDisabled={!isAdmin}
-                  colorScheme="blue"
-                  onClick={() => setIsAddModalOpen(true)}
-                  width={{ base: "100%", md: "auto" }}
-                >
-                  Add Batch
-                </Button>
-              </Flex>
-            </Center>
-            <Table
-              {...getTableProps()}
-              wordBreak={{ base: "normal", lg: "break-word" }}
-              background={baseColor}
-              colorScheme="customBaseColorScheme"
-              size="sm"
-            >
-              <Thead>
-                {headerGroups.map((headerGroup, index) => (
-                  <Tr {...headerGroup.getHeaderGroupProps()} key={index}>
-                    {headerGroup.headers.map(column => (
-                      <Th {...column.getHeaderProps(column.getSortByToggleProps())} key={column.id} whiteSpace="nowrap">
-                        {column.render("Header")}
-                        <chakra.span key={`span-${index}`} pl="4">
-                          {column.isSorted ? (
-                            column.isSortedDesc ? (
-                              <TriangleDownIcon aria-label="sorted descending" />
-                            ) : (
-                              <TriangleUpIcon aria-label="sorted ascending" />
-                            )
-                          ) : null}
-                        </chakra.span>
-                      </Th>
+            </Flex>
+          </Center>
+          <Table
+            {...getTableProps()}
+            wordBreak={{ base: "normal", lg: "break-word" }}
+            background={baseColor}
+            colorScheme="customBaseColorScheme"
+            size="sm"
+          >
+            <Thead>
+              {headerGroups.map((headerGroup, index) => (
+                <Tr {...headerGroup.getHeaderGroupProps()} key={index}>
+                  {headerGroup.headers.map(column => (
+                    <Th {...column.getHeaderProps(column.getSortByToggleProps())} key={column.id} whiteSpace="nowrap">
+                      {column.render("Header")}
+                      <chakra.span key={`span-${index}`} pl="4">
+                        {column.isSorted ? (
+                          column.isSortedDesc ? (
+                            <TriangleDownIcon aria-label="sorted descending" />
+                          ) : (
+                            <TriangleUpIcon aria-label="sorted ascending" />
+                          )
+                        ) : null}
+                      </chakra.span>
+                    </Th>
+                  ))}
+                </Tr>
+              ))}
+            </Thead>
+            <Tbody {...getTableBodyProps()}>
+              {page.map(row => {
+                prepareRow(row);
+                return (
+                  <Tr {...row.getRowProps()} key={row.id}>
+                    {row.cells.map(cell => (
+                      <Td {...cell.getCellProps()} key={cell.column.id}>
+                        {cell.render("Cell")}
+                      </Td>
                     ))}
                   </Tr>
+                );
+              })}
+            </Tbody>
+          </Table>
+
+          <Center mt={4}>
+            <ButtonGroup>
+              <Button onClick={() => gotoPage(0)} disabled={!canPreviousPage}>
+                {"<<"}
+              </Button>
+              <Button onClick={() => previousPage()} disabled={!canPreviousPage}>
+                {"<"}
+              </Button>
+              <Button onClick={() => nextPage()} disabled={!canNextPage}>
+                {">"}
+              </Button>
+              <Button onClick={() => gotoPage(pageCount - 1)} disabled={!canNextPage}>
+                {">>"}
+              </Button>
+            </ButtonGroup>
+          </Center>
+          <Center mt={4}>
+            <Text mr={4}>
+              Page{" "}
+              <strong>
+                {pageIndex + 1} of {pageOptions.length}
+              </strong>{" "}
+            </Text>
+            <Box>
+              <Select
+                isFullWidth={false}
+                value={pageSize}
+                onChange={e => {
+                  setPageSize(Number(e.target.value));
+                }}
+              >
+                {[25, 50, 100].map(pageSizeOption => (
+                  <option key={pageSizeOption} value={pageSizeOption}>
+                    Show {pageSizeOption}
+                  </option>
                 ))}
-              </Thead>
-              <Tbody {...getTableBodyProps()}>
-                {page.map(row => {
-                  prepareRow(row);
-                  return (
-                    <Tr {...row.getRowProps()} key={row.id}>
-                      {row.cells.map(cell => (
-                        <Td {...cell.getCellProps()} key={cell.column.id}>
-                          {cell.render("Cell")}
-                        </Td>
-                      ))}
-                    </Tr>
-                  );
-                })}
-              </Tbody>
-            </Table>
-            <Center mt={4}>
-              <ButtonGroup>
-                <Button onClick={() => gotoPage(0)} disabled={!canPreviousPage}>
-                  {"<<"}
-                </Button>
-                <Button onClick={() => previousPage()} disabled={!canPreviousPage}>
-                  {"<"}
-                </Button>
-                <Button onClick={() => nextPage()} disabled={!canNextPage}>
-                  {">"}
-                </Button>
-                <Button onClick={() => gotoPage(pageCount - 1)} disabled={!canNextPage}>
-                  {">>"}
-                </Button>
-              </ButtonGroup>
-            </Center>
-            <Center mt={4}>
-              <Text mr={4}>
-                Page{" "}
-                <strong>
-                  {pageIndex + 1} of {pageOptions.length}
-                </strong>{" "}
-              </Text>
-              <Box>
-                <Select
-                  isFullWidth={false}
-                  value={pageSize}
-                  onChange={e => {
-                    setPageSize(Number(e.target.value));
-                  }}
-                >
-                  {[25, 50, 100].map(pageSizeOption => (
-                    <option key={pageSizeOption} value={pageSizeOption}>
-                      Show {pageSizeOption}
-                    </option>
-                  ))}
-                </Select>
-              </Box>
-            </Center>
-          </Box>
-        )}
-      </Container>
-
-      {/* Add Batch Modal */}
-      <BatchCrudFormModal
-        mainnetProvider={mainnetProvider}
-        isOpen={isAddModalOpen}
-        onClose={() => setIsAddModalOpen(false)}
-        onUpdate={() => {
-          setIsAddModalOpen(false);
-          fetchBatches();
-        }}
-      />
-
-      {/* Edit Batch Modal */}
-      <BatchCrudFormModal
-        mainnetProvider={mainnetProvider}
-        batch={selectedBatch}
-        isOpen={isEditModalOpen}
-        onClose={() => setIsEditModalOpen(false)}
-        onUpdate={() => {
-          setIsEditModalOpen(false);
-          fetchBatches();
-        }}
-      />
-    </>
+              </Select>
+            </Box>
+          </Center>
+        </Box>
+      )}
+    </Container>
   );
 }
